exe speed optimization
replaced the std::vector<u8> memory system with one using raw void pointers.
added functions to create aligned memory arrays (currently 64-byte aligned in case mm512 intrinsics are added later).

fixed the cpu debug string writer.
added ostream flag preservation to all debug string writers.

memory access during execution is now done directly via pointer accesses.

added templated forms for raw memory access (pick a size at compile time and it'll fetch/store that to/from the u64& arg).
these are a boon for efficiency, as it avoids having to perform a switch in the (very common) case of having constant size components.

added a parity lookup table that is used for setting PF.
since a ton of (very common) instructions modify PF based on parity, this speeds up execution a lot.

simplified some of the flag computations (e.g. OF in ADD/SUB/CMP now use fancier bit twiddling to work their magic).
broke up a = b = c ... etc. flag assignments since assignment from the flag wrapper is slightly slower.

replaced all instances of 1ul (from C# impl) with (u64)1 - other ul literals left over from C# may cause problems since ul in C++ is frequently 32-bit.
these should be replaced with explicit use of u8/16/32/64 (or equivalent signed variants).

removed ExeTypes.cpp (now done inline in the header).
fixed some CPU/VPU sizecode wrappers not returning themselves on assignment (or returning a reference to themselves - which are already references).

removed utility functions for reading/writing c-strings (only used in the interpreter anyway).
removed a bunch of encoding functions from Utility.cpp and made them inlined in the header (and also constexpr where possible).

+++ AsmArgs.cpp
@@ -1878,7 +1878,7 @@ bool AssembleArgs::TryProcessFPUBinaryOp(OPCode op, bool integral, bool pop)
 		// if b is st(0) (do this one first since it handles the pop form)
 		if (b == 0)
 		{
-			if (!TryAppendVal(1, (a << 4) | (pop ? 2 : 1ul))) return false;
+			if (!TryAppendVal(1, (a << 4) | (pop ? 2 : (u64)1))) return false;
 		}
 		// if a is st(0)
 		else if (a == 0)
@@ -2108,7 +2108,7 @@ bool AssembleArgs::TryProcessFCOM(OPCode op, bool integral, bool pop, bool pop2,
 		// first arg must be st0
 		if (reg_a != 0) { res = {AssembleError::UsageError, "line " + tostr(line) + ": First operand must be ST(0)"};   return false; }
 
-		if (!TryAppendVal(1, (unordered ? 128 : 0ul) | (reg_b << 4) | (pop ? 12 : 11ul))) return false;
+		if (!TryAppendVal(1, (unordered ? 128 : 0ul) | (reg_b << 4) | (pop ? 12 : 11))) return false;
 	}
 	else { res = {AssembleError::ArgCount, "line " + tostr(line) + ": Too many operands"};   return false; }
